# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.DeviceActivity:
    properties:
      action:
        type: string
      deviceName:
        type: string
      gridName:
        type: string
      headers:
        description: Store as JSON string
        type: string
      id:
        type: integer
      sourceIP:
        type: string
      timestamp:
        type: string
      uniqueId:
        type: string
    type: object
  models.UsageStats:
    properties:
      endpoint:
        type: string
      id:
        type: string
      method:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: An API for tracking device activities and usage statistics.
  title: Activity API
  version: "1.0"
paths:
  /activities:
    get:
      description: Retrieves all recorded device activities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceActivity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all activities
      tags:
      - activities
    post:
      consumes:
      - application/json
      description: Records a new device activity with headers
      parameters:
      - description: Activity Data
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/models.DeviceActivity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DeviceActivity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new activity
      tags:
      - activities
  /activities/{id}:
    delete:
      description: Deletes a specific activity by ID
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an activity
      tags:
      - activities
  /activities/device/{device}:
    get:
      description: Retrieves activities for a specific device
      parameters:
      - description: Device Name
        in: path
        name: device
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceActivity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get activities by device
      tags:
      - activities
  /activities/grid/{grid}:
    get:
      description: Retrieves activities for a specific grid
      parameters:
      - description: Grid Name
        in: path
        name: grid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceActivity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get activities by grid
      tags:
      - activities
  /health:
    get:
      consumes:
      - application/json
      description: Returns OK if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - health
  /stats:
    get:
      description: Retrieves all usage statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UsageStats'
            type: array
      summary: Get all statistics
      tags:
      - stats
    post:
      consumes:
      - application/json
      description: Records new usage statistics
      parameters:
      - description: Stats Data
        in: body
        name: stats
        required: true
        schema:
          $ref: '#/definitions/models.UsageStats'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UsageStats'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create usage statistics
      tags:
      - stats
  /stats/{id}:
    delete:
      description: Deletes specific statistics by ID
      parameters:
      - description: Stats ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete statistics
      tags:
      - stats
  /stats/endpoints/{endpoint}:
    delete:
      description: Deletes all statistics for a specific endpoint
      parameters:
      - description: Endpoint Path
        in: path
        name: endpoint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete statistics by endpoint
      tags:
      - stats
    get:
      description: Retrieves statistics for a specific endpoint
      parameters:
      - description: Endpoint Path
        in: path
        name: endpoint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UsageStats'
            type: array
      summary: Get statistics by endpoint
      tags:
      - stats
swagger: "2.0"